macro_tools : evolve props
