+++ module/core/derive_tools_meta/src/derive/from.rs
@@ -1,5 +1,19 @@
 use super::*;
-use macro_tools::{ attr, diag, generic_params, item_struct, struct_like::StructLike, Result };
+use macro_tools::
+{
+  attr,
+  diag,
+  generic_params,
+  item_struct,
+  struct_like::StructLike,
+  Result,
+  AttributeComponent,
+  AttributePropertyComponent,
+  AttributePropertyBoolean,
+  // AttributePropertyEnabled,
+};
+
+use former_types::ComponentAssign;
 
 // xxx2 : get complete From for enums
 
@@ -88,9 +102,9 @@ pub fn from( input : proc_macro::TokenStream ) -> Result< proc_macro2::TokenStre
         .or_insert( 1 );
       });
 
-      let variants = item.variants.iter().map( | variant |
+      let variants_result : Result< Vec< proc_macro2::TokenStream > > = item.variants.iter().map( | variant |
       {
-        if map[ &variant.fields.to_token_stream().to_string() ] <= 1
+        if map[ & variant.fields.to_token_stream().to_string() ] <= 1
         {
           variant_generate
           (
@@ -103,9 +117,12 @@ pub fn from( input : proc_macro::TokenStream ) -> Result< proc_macro2::TokenStre
         }
         else
         {
-          qt!{}
+          Ok( qt!{} )
         }
-      });
+      }).collect();
+
+      let variants = variants_result?;
+
       qt!
       {
         #( #variants )*
@@ -131,14 +148,20 @@ fn variant_generate
   generics_where: &syn::punctuated::Punctuated< syn::WherePredicate, syn::token::Comma >,
   variant : &syn::Variant,
 )
--> proc_macro2::TokenStream
+-> Result< proc_macro2::TokenStream >
 {
   let variant_name = &variant.ident;
   let fields = &variant.fields;
+  let attrs = FieldAttributes::from_attrs( variant.attrs.iter() )?;
+
+  if !attrs.config.enabled.value( true )
+  {
+    return Ok( qt!{} )
+  }
 
   if fields.len() <= 0
   {
-    return qt!{}
+    return Ok( qt!{} )
   }
 
   let ( args, use_src ) = if fields.len() == 1
@@ -163,20 +186,23 @@ fn variant_generate
     )
   };
 
-  qt!
-  {
-    #[ automatically_derived ]
-    impl< #generics_impl > From< #args > for #item_name< #generics_ty >
-    where
-      #generics_where
+  Ok
+  (
+    qt!
     {
-      #[ inline ]
-      fn from( src : #args ) -> Self
+      #[ automatically_derived ]
+      impl< #generics_impl > From< #args > for #item_name< #generics_ty >
+      where
+        #generics_where
       {
-        Self::#variant_name( #use_src )
+        #[ inline ]
+        fn from( src : #args ) -> Self
+        {
+          Self::#variant_name( #use_src )
+        }
       }
     }
-  }
+  )
 
 }
 
@@ -357,28 +383,12 @@ fn generate_unit
 /// Attributes of a field / variant
 ///
 
-// xxx
-use macro_tools::
-{
-  attr,
-  syn_err,
-  return_syn_err,
-  qt,
-  Result,
-  AttributeComponent,
-  AttributePropertyComponent,
-  AttributePropertyBoolean,
-  AttributePropertyEnabled,
-};
-
-use former_types::{ ComponentAssign };
-
 /// Represents the attributes of a struct. Aggregates all its attributes.
 #[ derive( Debug, Default ) ]
 pub struct FieldAttributes
 {
   /// Attribute for customizing the mutation process.
-  pub from : AttributeFrom,
+  pub config : AttributeFrom,
 }
 
 impl FieldAttributes
@@ -470,6 +480,7 @@ impl FieldAttributes
 //   }
 //
 // }
+// xxx : clean
 
 #[ derive( Debug, Default ) ]
 pub struct AttributeFrom
@@ -511,7 +522,18 @@ where
   #[ inline( always ) ]
   fn assign( &mut self, component : IntoT )
   {
-    self.from = component.into();
+    self.config = component.into();
+  }
+}
+
+impl< IntoT > ComponentAssign< AttributePropertyEnabled, IntoT > for AttributeFrom
+where
+  IntoT : Into< AttributePropertyEnabled >,
+{
+  #[ inline( always ) ]
+  fn assign( &mut self, component : IntoT )
+  {
+    self.enabled = component.into();
   }
 }
 
@@ -526,17 +548,6 @@ where
   }
 }
 
-impl< IntoT > ComponentAssign< AttributePropertyEnabled, IntoT > for AttributeFrom
-where
-  IntoT : Into< AttributePropertyEnabled >,
-{
-  #[ inline( always ) ]
-  fn assign( &mut self, component : IntoT )
-  {
-    self.custom = component.into();
-  }
-}
-
 impl syn::parse::Parse for AttributeFrom
 {
   fn parse( input : syn::parse::ParseStream< '_ > ) -> syn::Result< Self >
@@ -548,8 +559,9 @@ impl syn::parse::Parse for AttributeFrom
       let known = const_format::concatcp!
       (
         "Known entries of attribute ", AttributeFrom::KEYWORD, " are : ",
-        AttributePropertyEnabled::KEYWORD,
-        ", ", AttributePropertyHint::KEYWORD,
+        AttributePropertyHint::KEYWORD,
+        ", ", AttributePropertyEnabled::KEYWORD_ON,
+        ", ", AttributePropertyEnabled::KEYWORD_OFF,
         ".",
       );
       syn_err!
@@ -569,10 +581,10 @@ impl syn::parse::Parse for AttributeFrom
       if lookahead.peek( syn::Ident )
       {
         let ident : syn::Ident = input.parse()?;
-
         match ident.to_string().as_str()
         {
-          AttributePropertyEnabled::KEYWORD => result.assign( AttributePropertyEnabled::parse( input )? ),
+          AttributePropertyEnabled::KEYWORD_ON => result.assign( AttributePropertyEnabled::from( true ) ),
+          AttributePropertyEnabled::KEYWORD_OFF => result.assign( AttributePropertyEnabled::from( false ) ),
           AttributePropertyHint::KEYWORD => result.assign( AttributePropertyHint::parse( input )? ),
           _ => return Err( error( &ident ) ),
         }
@@ -616,180 +628,6 @@ pub type AttributePropertyHint = AttributePropertyBoolean< AttributePropertyHint
 #[ derive( Debug, Default, Clone, Copy ) ]
 pub struct AttributePropertyEnabledMarker;
 
-impl AttributePropertyComponent for AttributePropertyEnabledMarker
-{
-  const KEYWORD : &'static str = "custom";
-}
-
 /// Indicates whether a custom code should be generated.
 /// Defaults to `false`, meaning no custom code is generated unless explicitly requested.
-pub type AttributePropertyEnabled = AttributePropertyEnabled< AttributePropertyEnabledMarker >;
-
-// pub struct FieldAttributes
-// {
-//   pub from : Option< AttributeFrom >,
-// }
-//
-// impl FieldAttributes
-// {
-//
-//   pub fn from_attrs< 'a >( attrs : impl Iterator< Item = &'a syn::Attribute > ) -> Result< Self >
-//   {
-//     let mut from : Option< AttributeFrom > = None;
-//
-//     for attr in attrs
-//     {
-//       let key_ident = attr.path().get_ident()
-//       .ok_or_else( || syn_err!( attr, "Expects an attribute of format #[ attribute( key1 = val1, key2 = val2 ) ], but got:\n  {}", qt!{ #attr } ) )?;
-//       let key_str = format!( "{}", key_ident );
-//
-//       if attr::is_standard( &key_str )
-//       {
-//         continue;
-//       }
-//
-//       // qqq : qqq for Anton : xxx : refactor field_attrs::FieldAttributes::from_attrs to make it similar to this function
-//       match key_str.as_ref()
-//       {
-//         AttributeFrom::KEYWORD =>
-//         {
-//           from.replace( AttributeFrom::from_meta( attr )? );
-//         }
-//         "debug" =>
-//         {
-//         }
-//         _ =>
-//         {
-//           return Err( syn_err!( attr, "Known field attirbutes are : `from`, `debug`.\nUnknown structure attribute : {}", qt!{ #attr } ) );
-//         }
-//       }
-//     }
-//
-//     Ok( FieldAttributes { from } )
-//   }
-//
-// }
-//
-//
-// ///
-// /// Attribute to hold parameters of forming for a specific field or variant.
-// /// For example to avoid code From generation for it.
-// ///
-// /// `#[ from( off, hint : true ) ]`
-// ///
-//
-// #[ derive( Default ) ]
-// pub struct AttributeFrom
-// {
-//   /// Specifies whether we should generate From implementation for the field.
-//   /// Can be altered using `on` and `off` attributes
-//   pub enabled : Option< bool >,
-//   /// Specifies whether to provide a sketch of generated From or not.
-//   /// Defaults to `false`, which means no hint is provided unless explicitly requested.
-//   pub hint : bool,
-// }
-//
-// impl AttributeFrom
-// {
-//
-//   const KEYWORD : &'static str = "from";
-//
-//   pub fn from_meta( attr : &syn::Attribute ) -> Result< Self >
-//   {
-//     match attr.meta
-//     {
-//       syn::Meta::List( ref meta_list ) =>
-//       {
-//         return syn::parse2::< AttributeFrom >( meta_list.tokens.clone() );
-//       },
-//       syn::Meta::Path( ref _path ) =>
-//       {
-//         return Ok( Default::default() )
-//       },
-//       _ => return_syn_err!( attr, "Expects an attribute of format #[ from( off ) ]
-// .\nGot: {}", qt!{ #attr } ),
-//     }
-//   }
-//
-// }
-//
-// impl syn::parse::Parse for AttributeFrom
-// {
-//   fn parse( input : syn::parse::ParseStream< '_ > ) -> syn::Result< Self >
-//   {
-//     let mut off : bool = false;
-//     let mut on : bool = false;
-//     let mut hint = false;
-//
-//     while !input.is_empty()
-//     {
-//       let lookahead = input.lookahead1();
-//       if lookahead.peek( syn::Ident )
-//       {
-//         let ident : syn::Ident = input.parse()?;
-//         // xxx : qqq for Anton : use match here and for all attributes -- done
-//         match ident.to_string().as_str()
-//         {
-//           "off" =>
-//           {
-//             input.parse::< syn::Token![ = ] >()?;
-//             let value : syn::LitBool = input.parse()?;
-//             off = value.value();
-//           },
-//           "on" =>
-//           {
-//             input.parse::< syn::Token![ = ] >()?;
-//             let value : syn::LitBool = input.parse()?;
-//             on = value.value();
-//           }
-//           "hint" =>
-//           {
-//             input.parse::< syn::Token![ = ] >()?;
-//             let value : syn::LitBool = input.parse()?;
-//             hint = value.value;
-//           }
-//           _ =>
-//           {
-//             return Err( syn::Error::new_spanned( &ident, format!( "Unexpected identifier '{}'. Expected 'on', 'off', or 'hint'. For example: `#[ from( off, hint : true ) ]`", ident ) ) );
-//           }
-//         }
-//       }
-//       else
-//       {
-//         return Err( syn::Error::new( input.span(), "Unexpected identifier '{}'. Expected 'on', 'off', or 'hint'. For example: `#[ from( off, hint : true ) ]`" ) );
-//       }
-//
-//     }
-//
-//     // xxx : move on / off logic into a helper
-//
-//     let mut enabled : Option< bool > = None;
-//
-//     if on && off
-//     {
-//       // return Err( syn_err!( input, "`on` and `off` are mutually exclusive .\nIllegal attribute usage : {}", qt!{ #input } ) )
-//       return Err( syn::Error::new( input.span(), "`on` and `off` are mutually exclusive .\nIllegal attribute usage" ) );
-//       // xxx : test
-//     }
-//
-//     if !on && !off
-//     {
-//       enabled = None;
-//     }
-//     else if on
-//     {
-//       enabled = Some( true )
-//     }
-//     else if off
-//     {
-//       enabled = Some( false )
-//     }
-//
-//     // Optional comma handling
-//     if input.peek( syn::Token![ , ] )
-//     {
-//       input.parse::< syn::Token![ , ] >()?;
-//     }
-//     Ok( Self { enabled, hint } )
-//   }
-// }
+pub type AttributePropertyEnabled = macro_tools::AttributePropertyEnabled< AttributePropertyEnabledMarker >;
