+++ module/core/former_meta/Cargo.toml
@@ -58,7 +58,7 @@ proc-macro = true
 
 [dependencies]
 macro_tools = { workspace = true } # qqq : optimize set of features
-former_types = { workspace = true, features = [ "types_component_assign" ] }
+former_types = { workspace = true, features = [ "enabled", "types_component_assign" ] }
 iter_tools = { workspace = true }
 convert_case = { version = "0.6.0", default-features = false, optional = true, features = [] }
 const_format = { version = "0.2.32" }
