+++ module/core/macro_tools/src/attr_prop/syn.rs
@@ -1,0 +1,94 @@
+//!
+//! Property of an attribute which simply wraps one of the standard `syn` types.
+//!
+
+use crate::*;
+
+///
+/// Property of an attribute which simply wraps one of the standard `syn` types.
+///
+
+#[ derive( Debug, Clone ) ]
+pub struct AttributePropertySyn< T, Marker >( T, ::core::marker::PhantomData< Marker > )
+where
+  T : syn::parse::Parse + quote::ToTokens;
+
+impl< T, Marker > AttributePropertySyn< T, Marker >
+where
+  T : syn::parse::Parse + quote::ToTokens,
+{
+  /// Just unwraps and returns the internal data.
+  #[ allow( dead_code ) ]
+  pub fn internal( self ) -> T
+  {
+    self.0
+  }
+
+  /// Returns a reference to the internal data.
+  #[ allow( dead_code ) ]
+  pub fn ref_internal( &self ) -> &T
+  {
+    &self.0
+  }
+}
+
+impl< T, Marker > AttributePropertyComponent for AttributePropertySyn< T, Marker >
+where
+  T : syn::parse::Parse + quote::ToTokens,
+  Marker : AttributePropertyComponent,
+{
+  const KEYWORD : &'static str = Marker::KEYWORD;
+}
+
+impl< T, Marker > syn::parse::Parse for AttributePropertySyn< T, Marker >
+where
+  T : syn::parse::Parse + quote::ToTokens,
+{
+  fn parse( input : syn::parse::ParseStream< '_ > ) -> syn::Result< Self >
+  {
+    input.parse::< syn::Token![ = ] >()?;
+    let value : T = input.parse()?;
+    Ok( value.into() )
+  }
+}
+
+impl< T, Marker > quote::ToTokens for AttributePropertySyn< T, Marker >
+where
+  T : syn::parse::Parse + quote::ToTokens,
+{
+  fn to_tokens( &self, tokens : &mut proc_macro2::TokenStream )
+  {
+    self.0.to_tokens( tokens );
+  }
+}
+
+impl< T, Marker > core::ops::Deref for AttributePropertySyn< T, Marker >
+where T : syn::parse::Parse + quote::ToTokens
+{
+  type Target = T;
+  #[ inline( always ) ]
+  fn deref( &self ) -> &T
+  {
+    &self.0
+  }
+}
+
+impl< T, Marker > AsRef< T > for AttributePropertySyn< T, Marker >
+where T : syn::parse::Parse + quote::ToTokens
+{
+  #[ inline( always ) ]
+  fn as_ref( &self ) -> &T
+  {
+    &self.0
+  }
+}
+
+impl< T, Marker > From< T > for AttributePropertySyn< T, Marker >
+where T : syn::parse::Parse + quote::ToTokens
+{
+  #[ inline( always ) ]
+  fn from( src : T ) -> Self
+  {
+    Self( src, Default::default() )
+  }
+}
