+++ module/core/macro_tools/tests/inc/attr_prop_test.rs
@@ -100,3 +100,14 @@ fn attr_prop_test()
   assert_eq!( parsed.debug.internal(), false );
 
 }
+
+#[ test ]
+fn attribute_property_enabled()
+{
+  // Test default value
+  let attr : AttributePropertyEnabled = Default::default();
+  assert_eq!( attr.internal(), None );
+  assert_eq!( attr.value( true ), true );
+  assert_eq!( attr.value( false ), false );
+
+}
